+++ src/font-awesome.js
@@ -59,6 +59,7 @@ import {
     faThumbtack,
     faTimes,
     faTrashAlt,
+    faUpload,
     faUser
 } from '@fortawesome/free-solid-svg-icons';
 
@@ -120,4 +121,5 @@ library.add(faTasks);
 library.add(faThumbtack);
 library.add(faTimes);
 library.add(faTrashAlt);
+library.add(faUpload);
 library.add(faUser);