+++ src/reducers/Reducers.js
@@ -4,6 +4,7 @@ import undoable, { excludeAction } from 'redux-undo';
 import App from 'reducers/App';
 import Notes from 'reducers/Notes';
 import Objects from 'reducers/Objects';
+import Publication from 'reducers/Publication';
 import Session from 'reducers/Session';
 import Settings from 'reducers/Settings';
 import Synchronization from 'reducers/Synchronization';
@@ -28,6 +29,7 @@ export default combineReducers({
     session: Session(),
     settings: Settings(),
     synchronization: Synchronization(),
+    publication: Publication(),
     tasks: undoable(reduceReducers([], Objects('tasks'), Tasks()), {
         limit: 5,
         filter: process.env.REACT_APP_MODE === 'electron' ? excludeAction(['SET_OBJECTS', 'CHANGE_ID', 'CLEAN_OBJECTS']) : () => false,